Minor layout tweaks to Notifications view. And it now fetches events from relays.
