+++ Nos/Service/Filter.swift
@@ -14,6 +14,7 @@ struct Filter: Hashable, Identifiable {
     let eventIDs: [HexadecimalString]
     let kinds: [EventKind]
     let eTags: [HexadecimalString]
+    let pTags: [HexadecimalString]
     let limit: Int?
     let since: Date?
     
@@ -26,6 +27,7 @@ struct Filter: Hashable, Identifiable {
         eventIDs: [HexadecimalString] = [],
         kinds: [EventKind] = [],
         eTags: [HexadecimalString] = [],
+        pTags: [HexadecimalString] = [],
         limit: Int? = nil,
         since: Date? = nil
     ) {
@@ -33,6 +35,7 @@ struct Filter: Hashable, Identifiable {
         self.eventIDs = eventIDs
         self.kinds = kinds.sorted(by: { $0.rawValue > $1.rawValue })
         self.eTags = eTags
+        self.pTags = pTags
         self.limit = limit
         self.since = since
     }
@@ -58,6 +61,10 @@ struct Filter: Hashable, Identifiable {
         
         if !eTags.isEmpty {
             filterDict["#e"] = eTags
+        }
+        
+        if !pTags.isEmpty {
+            filterDict["#p"] = pTags
         }
         
         if let since {
