+++ Nos/Views/SelectableText.swift
@@ -26,7 +26,7 @@ struct SelectableText: UIViewRepresentable {
     init(_ string: String) {
         self.attributedText = AttributedString(string)
         attributedText.setAttributes(AttributeContainer([
-            .font: UIFont.clarity!,
+            .font: UIFont.clarity,
             .foregroundColor: UIColor.primaryTxt,
         ]))
     }
